requirements.txt
Django==5.1
drf-spectacular==0.27.2
django_extensions==3.2.3
django-cors-headers==4.5.0
pytest==7.3.1
pytest-django==4.5.2
pillow==10.3.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djrestframework-simplejwt==1.22.0